Mostly working refactor
